move delete to action creator
