+++ client/src/components/Sidebar.jsx
@@ -1,7 +1,29 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import { useHistory } from "react-router";
+import Swal from "sweetalert2";
 
 export default function Sidebar() {
+  const history = useHistory();
+
+  function toLogout() {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, Logout",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        history.push("/login");
+        Swal.fire("Logout!", "See You Again!", "success");
+      }
+    });
+  }
+
   return (
     <div>
       <h1>navbar</h1>
@@ -164,7 +186,7 @@ export default function Sidebar() {
             </li>
             <li>
               <Link
-                href="#"
+                onClick={toLogout}
                 className="
                 relative
                 flex flex-row
@@ -183,136 +205,22 @@ export default function Sidebar() {
               >
                 <span className="inline-flex justify-center items-center ml-4">
                   <svg
-                    className="w-5 h-5"
+                    class="w-5 h-5"
                     fill="none"
                     stroke="currentColor"
                     viewBox="0 0 24 24"
                     xmlns="http://www.w3.org/2000/svg"
                   >
                     <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M20 13V6a2 2 0 00-2-2H6a2 2 0 00-2 2v7m16 0v5a2 2 0 01-2 2H6a2 2 0 01-2-2v-5m16 0h-2.586a1 1 0 00-.707.293l-2.414 2.414a1 1 0 01-.707.293h-3.172a1 1 0 01-.707-.293l-2.414-2.414A1 1 0 006.586 13H4"
+                      stroke-linecap="round"
+                      stroke-linejoin="round"
+                      stroke-width="2"
+                      d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"
                     ></path>
                   </svg>
                 </span>
                 <span className="ml-2 text-sm tracking-wide truncate">
-                  Board
-                </span>
-                <span
-                  className="
-                  hidden
-                  md:block
-                  px-2
-                  py-0.5
-                  ml-auto
-                  text-xs
-                  font-medium
-                  tracking-wide
-                  text-blue-500
-                  bg-indigo-50
-                  rounded-full
-                "
-                >
-                  New
-                </span>
-              </Link>
-            </li>
-            <li>
-              <Link
-                href="#"
-                className="
-                relative
-                flex flex-row
-                items-center
-                h-11
-                focus:outline-none
-                hover:bg-blue-800
-                dark:hover:bg-gray-600
-                text-white-600
-                hover:text-white-800
-                border-l-4 border-transparent
-                hover:border-blue-500
-                dark:hover:border-gray-800
-                pr-6
-              "
-              >
-                <span className="inline-flex justify-center items-center ml-4">
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M7 8h10M7 12h4m1 8l-4-4H5a2 2 0 01-2-2V6a2 2 0 012-2h14a2 2 0 012 2v8a2 2 0 01-2 2h-3l-4 4z"
-                    ></path>
-                  </svg>
-                </span>
-                <span className="ml-2 text-sm tracking-wide truncate">
-                  Messages
-                </span>
-              </Link>
-            </li>
-            <li>
-              <Link
-                href="#"
-                className="
-                relative
-                flex flex-row
-                items-center
-                h-11
-                focus:outline-none
-                hover:bg-blue-800
-                dark:hover:bg-gray-600
-                text-white-600
-                hover:text-white-800
-                border-l-4 border-transparent
-                hover:border-blue-500
-                dark:hover:border-gray-800
-                pr-6
-              "
-              >
-                <span className="inline-flex justify-center items-center ml-4">
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M15 17h5l-1.405-1.405A2.032 2.032 0 0118 14.158V11a6.002 6.002 0 00-4-5.659V5a2 2 0 10-4 0v.341C7.67 6.165 6 8.388 6 11v3.159c0 .538-.214 1.055-.595 1.436L4 17h5m6 0v1a3 3 0 11-6 0v-1m6 0H9"
-                    ></path>
-                  </svg>
-                </span>
-                <span className="ml-2 text-sm tracking-wide truncate">
-                  Notifications
-                </span>
-                <span
-                  className="
-                  hidden
-                  md:block
-                  px-2
-                  py-0.5
-                  ml-auto
-                  text-xs
-                  font-medium
-                  tracking-wide
-                  text-red-500
-                  bg-red-50
-                  rounded-full
-                "
-                >
-                  1.2k
+                  Logout
                 </span>
               </Link>
             </li>
@@ -322,92 +230,6 @@ export default function Sidebar() {
                   Settings
                 </div>
               </div>
-            </li>
-            <li>
-              <Link
-                href="#"
-                className="
-                relative
-                flex flex-row
-                items-center
-                h-11
-                focus:outline-none
-                hover:bg-blue-800
-                dark:hover:bg-gray-600
-                text-white-600
-                hover:text-white-800
-                border-l-4 border-transparent
-                hover:border-blue-500
-                dark:hover:border-gray-800
-                pr-6
-              "
-              >
-                <span className="inline-flex justify-center items-center ml-4">
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"
-                    ></path>
-                  </svg>
-                </span>
-                <span className="ml-2 text-sm tracking-wide truncate">
-                  Profile
-                </span>
-              </Link>
-            </li>
-            <li>
-              <Link
-                href="#"
-                className="
-                relative
-                flex flex-row
-                items-center
-                h-11
-                focus:outline-none
-                hover:bg-blue-800
-                dark:hover:bg-gray-600
-                text-white-600
-                hover:text-white-800
-                border-l-4 border-transparent
-                hover:border-blue-500
-                dark:hover:border-gray-800
-                pr-6
-              "
-              >
-                <span className="inline-flex justify-center items-center ml-4">
-                  <svg
-                    className="w-5 h-5"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z"
-                    ></path>
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth="2"
-                      d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
-                    ></path>
-                  </svg>
-                </span>
-                <span className="ml-2 text-sm tracking-wide truncate">
-                  Settings
-                </span>
-              </Link>
             </li>
           </ul>
           <p className="mb-14 px-5 py-3 hidden md:block text-center text-xs">
