+++ client/src/pages/CemetaryKeeper.jsx
@@ -27,11 +27,6 @@ export default function CemetaryKeeper() {
     return <h1>ada error ...</h1>;
   }
 
-  function toAddForm() {
-    // console.log("aku di klik yeay/");
-    history.push("/cemetaryKeeper/add");
-  }
-
   return (
     <div className="h-full ml-14 mt-14 mb-10 md:ml-64 ">
       <h1 className="title-font font-semibold text-4xl text-center text-blue-600 mt-1 rounded-sm">
@@ -40,7 +35,7 @@ export default function CemetaryKeeper() {
       <div className="mx-12">
         <button
           className="hover:bg-blue-200 hover:text-blue-800 group flex items-center rounded-md bg-blue-100 text-blue-600 text-sm font-medium px-4 py-2"
-          onClick={toAddForm}
+          // onClick={toAddForm}
         >
           <svg
             className="group-hover:text-blue-600 text-blue-500 mr-2"
