+++ client/src/pages/Dashboard.jsx
@@ -90,7 +90,7 @@ export default function Dashboard() {
             </svg>
           </div>
           <div className="text-right">
-            <p className="text-2xl">557</p>
+            <p className="text-2xl">{transaction.length}</p>
             <p>Orders</p>
           </div>
         </div>
@@ -250,164 +250,6 @@ export default function Dashboard() {
                 return <BodyTableTransaction key={item._id} item={item} />;
               })}
             </table>
-          </div>
-          <div
-            className="
-              grid
-              px-4
-              py-3
-              text-xs
-              font-semibold
-              tracking-wide
-              text-gray-500
-              uppercase
-              border-t
-              dark:border-gray-700
-              bg-gray-50
-              sm:grid-cols-9
-              dark:text-gray-400 dark:bg-gray-800
-            "
-          >
-            <span className="flex items-center col-span-3">
-              Showing 21-30 of 100
-            </span>
-            <span className="col-span-2"></span>
-            {/*===PAGINATION===*/}
-            <span className="flex col-span-4 mt-2 sm:mt-auto sm:justify-end">
-              <nav aria-label="Table navigation">
-                <ul className="inline-flex items-center">
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md rounded-l-lg
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                      aria-label="Previous"
-                    >
-                      <svg
-                        aria-hidden="true"
-                        className="w-4 h-4 fill-current"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
-                          clipRule="evenodd"
-                          fillRule="evenodd"
-                        ></path>
-                      </svg>
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      1
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      2
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        text-white
-                        dark:text-gray-800
-                        transition-colors
-                        duration-150
-                        bg-blue-600
-                        dark:bg-gray-100
-                        border border-r-0 border-blue-600
-                        dark:border-gray-100
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      3
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      4
-                    </button>
-                  </li>
-                  <li>
-                    <span className="px-3 py-1">...</span>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      8
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                    >
-                      9
-                    </button>
-                  </li>
-                  <li>
-                    <button
-                      className="
-                        px-3
-                        py-1
-                        rounded-md rounded-r-lg
-                        focus:outline-none focus:shadow-outline-purple
-                      "
-                      aria-label="Next"
-                    >
-                      <svg
-                        className="w-4 h-4 fill-current"
-                        aria-hidden="true"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                          clipRule="evenodd"
-                          fillRule="evenodd"
-                        ></path>
-                      </svg>
-                    </button>
-                  </li>
-                </ul>
-              </nav>
-            </span>
           </div>
         </div>
       </div>
